requirements.txt
attr==0.3.2
imageio==2.15.0
matplotlib==3.3.4
mmcv==2.1.0
mmcv_full==1.7.0
mmseg==1.3.0
mmsegmentation==0.30.0
numpy==1.19.5
opencv_python==4.7.0.72
Pillow==10.2.0
PyYAML==6.0.1
scikit_image==0.17.2
skimage==0.0
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
timm==0.6.12
torch==1.10.0+cu113
torchsummary==1.5.1
torchvision==0.11.0+cu113
tqdm==4.64.1
